+++ collect_changelogs.py
@@ -36,7 +36,7 @@ from packaging import version
 from jinja2 import Template
 
 def download_zip(target_filename):
-    req = requests.get("https://github.com/opnsense/changelog/archive/master.zip", stream=True)
+    req = requests.get("https://github.com/reticen8/changelog/archive/master.zip", stream=True)
     if req.status_code == 200:
         req.raw.decode_content = True
         with open(target_filename, 'wb') as f_out:
@@ -73,7 +73,7 @@ def parse_change_log(payload, this_version, links):
         elif first_line is False and line.strip() != "":
             # strip tag line
             first_line = idx
-            if line.find('OPNsense') > -1:
+            if line.find('Reticen8') > -1:
                 content_line = line
         elif line == '--':
             # chop tagine