+++ source/conf.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 #
-# OPNsense documentation build configuration file, created by
+# Reticen8 documentation build configuration file, created by
 # sphinx-quickstart on Tue Jan 19 16:15:33 2016.
 #
 # This file is execfile()d with the current directory set to its
@@ -71,9 +71,9 @@ source_suffix = '.rst'
 master_doc = 'index'
 
 # General information about the project.
-project = u'OPNsense'
+project = u'Reticen8'
 copyright = u'2016-%d, Reticen8 B.V' % datetime.datetime.now().year
-author = u'Reticen8 B.V.'
+author = u'Reticen8 Technologies'
 
 # The version info for the project you're documenting, acts as replacement for
 # |version| and |release|, also used in various other places throughout the
@@ -135,7 +135,7 @@ todo_include_todos = True
 # a list of builtin themes.
 # html_theme = 'alabaster'
 html_theme = "sphinx_rtd_theme"
-#html_style = 'css/opnsense.css'
+#html_style = 'css/reticen8.css'
 
 # Theme options are theme-specific and customize the look and feel of a theme
 # further.  For a list of options available for each theme, see the
@@ -237,7 +237,7 @@ html_show_sourcelink = False
 #html_search_scorer = 'scorer.js'
 
 # Output file base name for HTML help builder.
-htmlhelp_basename = 'OPNsensedoc'
+htmlhelp_basename = 'Reticen8doc'
 
 # -- Options for LaTeX output ---------------------------------------------
 
@@ -259,8 +259,8 @@ latex_elements = {
 # (source start file, target name, title,
 #  author, documentclass [howto, manual, or own class]).
 latex_documents = [
-    (master_doc, 'OPNsense.tex', u'OPNsense Documentation',
-     u'Reticen8 B.V.', 'manual'),
+    (master_doc, 'Reticen8.tex', u'Reticen8 Documentation',
+     u'Reticen8 Technologies', 'manual'),
 ]
 
 # The name of an image file (relative to this directory) to place at the top of
@@ -289,7 +289,7 @@ latex_documents = [
 # One entry per manual page. List of tuples
 # (source start file, name, description, authors, manual section).
 man_pages = [
-    (master_doc, 'opnsense', u'OPNsense Documentation',
+    (master_doc, 'reticen8', u'Reticen8 Documentation',
      [author], 1)
 ]
 
@@ -303,8 +303,8 @@ man_pages = [
 # (source start file, target name, title, author,
 #  dir menu entry, description, category)
 texinfo_documents = [
-    (master_doc, 'OPNsense', u'',
-     author, 'OPNsense', 'High-end security made easy.',
+    (master_doc, 'Reticen8', u'',
+     author, 'Reticen8', 'High-end security made easy.',
      'Miscellaneous'),
 ]